Rethinking
The Render optimisations are a fine game to play, but i i have run into
some problems with it. the height of the Glyphs are not set correctly,
wich means i get out of bounds errors on characters like y.

After contemplating a bit and reminding myself that premature
optimisatios are the root of all evil, i have decided to stop the glyph
drawing and continue with sorting the input and displaying it on screen.
Even if it is slow. Once this stands, i can optimise how i want.